+++ metric/instrument_test.go
@@ -129,9 +129,9 @@ func TestWithAttributesConcurrentSafe(*testing.T) {
 	wg.Wait()
 }
 
-func TestResettableOptions(t *testing.T) {
-	type resettable interface {
-		Reset(attribute.Set)
+func TestSettableOptions(t *testing.T) {
+	type settable interface {
+		Set(attribute.Set)
 	}
 
 	aliceAttr := attribute.String("user", "Alice")
@@ -141,20 +141,20 @@ func TestResettableOptions(t *testing.T) {
 
 	t.Run("WithAttributeSet", func(t *testing.T) {
 		opt := WithAttributeSet(alice)
-		r, ok := opt.(resettable)
-		require.True(t, ok, "WithAttributeSet option does not implement resettable")
+		r, ok := opt.(settable)
+		require.True(t, ok, "WithAttributeSet option does not implement settable")
 
-		r.Reset(bob)
+		r.Set(bob)
 		c := NewAddConfig([]AddOption{opt.(AddOption)})
 		assert.Equal(t, bob, c.Attributes())
 	})
 
 	t.Run("WithAttributes", func(t *testing.T) {
 		opt := WithAttributes(aliceAttr)
-		r, ok := opt.(resettable)
-		require.True(t, ok, "WithAttributes option does not implement resettable")
+		r, ok := opt.(settable)
+		require.True(t, ok, "WithAttributes option does not implement settable")
 
-		r.Reset(bob)
+		r.Set(bob)
 		c := NewAddConfig([]AddOption{opt.(AddOption)})
 		assert.Equal(t, bob, c.Attributes())
 	})